feature: re-arch package core and impl mod in each file respective
