+++ src/constants.rs
@@ -1,0 +1,2 @@
+pub const HOST_REGEX: &str = r"^(localhost|127\.0\.0\.1)$";
+pub const PORT_REGEX: &str = r"^((6553[0-5])|(655[0-2][0-9])|(65[0-4][0-9]{2})|(6[0-4][0-9]{3})|([1-5][0-9]{4})|([0-5]{0,5})|([0-9]{1,4}))$";
