+++ tests/lib_spec.rs
@@ -1,12 +1,12 @@
 #[cfg(test)]
 mod params {
-  use core::params::Params;
+  use core::params;
 
   #[test]
   fn is_empty() {
     let host = "127.0.0.1";
     let port = 4000;
-    let result = Params::is_empty(host, port);
+    let result = params::is_empty(host, port);
 
     assert_eq!(result, false);
   }
@@ -14,7 +14,7 @@ mod params {
   #[test]
   fn is_valid_local_host() {
     let host = "127.0.0.1";
-    let result = Params::is_valid_local_host(host);
+    let result = params::is_valid_local_host(host);
 
     assert_eq!(result, true);
   }
@@ -22,7 +22,7 @@ mod params {
   #[test]
   fn is_valid_port() {
     let port = 8080;
-    let result = Params::is_valid_port(port);
+    let result = params::is_valid_port(port);
 
     assert_eq!(result, true);
   }
